Add the fail method and improve stack traces
The fail() method is not in the spec in the
https://github.com/google/skylark repository
but it is necessary to support the conformance
case.

It is true that different implementation might
want to change the behavior of fail() but it
can be done by overloading it in the
environment.

Stack traces where used mostly for recursion
detection but they are also used for failure
information now so store a friendly message for each stack call to show
to the user.