Use weak reference when adding a structure recursively.
This might lead to pretty expensive operation because we go through
the structure recursively when adding it to some other container
but it guarantee that we don't create unfreeable memory by creating
a Rc<> cycle.

Every time we encounter a weak reference we also do special tests
for comparisons and string conversion in order to avoid overflowing
the stack.