Forbid recursive data structure
Recursive data structure are allowed by the go implementation but they
make the language Turing complete which should not be allowed.

Furthermore, there is a bug that cannot be solved with weak pointers: a
cycle might be removed without the weak pointer being upgraded, this
would create a structure that can be released even though some weak
pointers are still kept.