Split the crate in 2
The main crate had a lot of transitive dependencies due to the REPL
so this change extract the REPL in a separate crate. This way
the main Starlark crate only have 9 non-build transitive dependencies:

$ cargo tree
starlark v0.2.0-pre (/usr/local/google/home/dmarting/git/skylark-rust/starlark)
├── codemap v0.1.1
├── codemap-diagnostic v0.1.0
│   ├── codemap v0.1.1 (*)
│   ├── isatty v0.1.9
│   │   ├── cfg-if v0.1.6
│   │   └── libc v0.2.45
│   └── term v0.4.6
├── lalrpop-util v0.16.2
└── linked-hash-map v0.5.1

This is not updating the documentation yet.

Issue #23