Value is not TypedValue
There is no way to implement things returning references to the
TypedValue from the Value, because reference lifetimes.

There is no way to add functions which return artbirary types to the
TypedValue trait, because that would require generics, which are not
object safe.

So we stop Value being a TypedValue, so that we will be able to add
functions returning references to TypedValue.