[refactor] Introduce a dialect domain concept.
Language dialect is an important domain concept that should
be explicitly called out in code. It makes it much easier to
document it and makes code using it easier to follow, since
instead of `true` and `false` we use `Dialect::Build` or `Dialect::Bzl`.
This can also come very handy in case some useful methods should
be available for it.