Add support for set collections, behind a feature (#45)
Pants allows the use of sets in BUILD files, so this is useful for experimenting with starlark in Pants.

Equality and ordering semantics are slightly weird here; sets preserve insertion order, but are considered equal ignoring order. Ordering comparison of sets is consistent with this (and consistent in general), but ill-defined; it's not based on number of elements, or anything like that, it just uses a combined hash of the contents so that hash is consistent with eq.

This also is slightly different from the dict implementation in that it doesn't rely on Value implementing Hash and instead uses a wrapper type so that callers are forced to eagerly handle unhashable types, rather than us panicing. I'd be interested in extending this to the dict type, if that's agree to be useful, too.

Authored-By: illicitonion <dawagner@gmail.com>