TypeValues type
"Caller environment" can only be used to revolve "type values" and
nothing more.

It might be convenient to access true caller environment, but it leads
to certain problems. Consider this situation. Module `a.bzl`
defines a function `foo` which calls function `bar`.

```
def bar(): ...

def foo(): bar()
```

and `BUILD` file imports `a.bzl` and calls `foo`:

```
foo()
```

`foo` function captures the context of `a.bzl` and inherits it in
locals (since #113).

However, when `foo` resolves type values, it must use type values
defined in the root caller (`BUILD` context, not captured).

Now function `foo` calls `bar`. Function `bar` also needs to resolve
type values.  These type values must again come from `BUILD` (because
captured context might not contain proper type values).

This environment passed to `bar` must be `BUILD` environment, but
not the local environment of `foo`.

Why the environment passed to `bar` must be `BUILD` environment?
Consider this scenario, `baz` is defined in `baz.bzl`:

```
def baz(unknown): unknown.append(10)
```

When `baz` is called from `BUILD` with object `foo` defined elsewhere,
`baz` should use the caller context, not local or captured context
to resolve `unknown.bar`. Because local (and captured) context might
not contain type values for `unknown`. But the function definition
is perfectly valid.

In theory in `foo` when calling `bar` we could merge the local
environment with the environment from `BUILD`, but:

* it is not practically needed for anything
* it complicates everything beyond the limit of understanding by human

`TypeValues` type is introduced which wraps `Environment`
but only provides access to type values. And `Environment` no longer
provides readable access to type values.

The good side is that change forced me to fix a couple of issues
with using an incorrect environment to find type values.

The bad side is that `dbg` function is no longer able to print
caller locals.

But the good side is that `dbg` was broken anyway after fix #113.