Make recursion counter thread-local
This practically changes nothing for `compare` where it is used
now, but it could be used in other operations which are potentially
recursive.

Simple example: this code example crashes Starlark REPL:

```
def f():
    a = []
    for i in range(100000):
        a = [a]
    return len(str(a))

print(f())
```

It should return a proper error instead. A thread-local counter can be
used to detect too deep recursion.

We could patch all functions signatures to accept `recursion` (or
`context`) parameter, however:

* thread-local is convenient here
* explicit parameters cannot be passed through external functions
  where signatures cannot be modified (like `Vec::eq`)

Similarly, we can implement thread-local allocation counters (to
avoid OOM) and instruction counters (to prevent programs which are
guaranteed to terminate in million years).