Use flat variable storage for nested scopes in local context
After this diff, comprehensions create a scope only during analysis,
but at runtime, comprehension locals are assigned different slots
in a flat variable space, as it happens in grown-up interpreters
or compilers.

Instead of a function scope, we have "local" scope which is created
for:
* function locals (including nested comprehension scopes)
* comprehension expressions in the module context

It's worth mentioning that in a local context (e. g. in a function
context) it is no longer possible to access a variable by name,
only by slot index (because name to slot index is no longer unique).

No slot optimization is performed at the moment, e. g. this program:

```
_ = [x for x in []]
_ = [x for x in []]
```

creates two slots for `x` variables in a function context. Should not
be a big issue, though.

This change significantly complicates analysis phase but simplifies
the runtime, which should be helpful to implement more optimizations
or GC.