Replace set constructor with flag
It's unlikely that someone would needs alternative `set` implementation,
same way as it's unlikely that someone would need alternative `dict`
implementation.

So replace dynamic `set_constructor` function with simple `set_literals`
flag.

Note `depset` is not an alternative implementation of sets, e. g.
it is not iterable, while `set` should be, and `depset` is initialized
with `order` parameter. So set literal should not be be used to
construct depsets.