Store * and + operator implementations in TypeValues
Certain operations like `+` or `*` cannot be defined on LHS.

For example to be able to embed Starlark in Bazel, implementation
should be able to override `+` for list and select types:

```
[] + select({...})
```

To solve this issue Python types define a pair of operators, for
plus it is `__add__` and `__radd__`. Former is called on LHS, and
if it is not defined, `__radd__` is used on RHS.

We could do the same, but alternatively we can store operator
implementations for each type pair in `TypeValues` object, this way
operators become symmetrical between LHS and RHS.

This diff implements storing `+` and `*` in `TypeValues`, the rest
operators (including unary operators) are still implemented as
virtual functions of `TypedValue`.