Optimize downcast_ref
Since we no longer depend on `RefCell`, we can add `ObjectRef::flat_map`
operation which maps a pointer to an option.

Before this diff downcasting did two virtual calls to `Any::type_id`
(in `Any::is` and in `Any::downcast_ref`), and now there is one.

Artificial integer addition benchmark (which does a lot of downcasting)
becomes about 5% faster.